mysql user creation works!

+++ mysql/providers/user.rb
@@ -20,13 +20,19 @@
 
 include Chef::Provider::Mysql::Base
 
-def validate
+action :validate do
   raise "Password required" unless @new_resource.password
+  Chef::Log.debug("Pre-validate options #{@new_resource.options.inspect}")
   @new_resource.host ||= '%'
+  @new_resource.credentials ||= {
+    :user => 'root',
+    :password => node['mysql']['server_root_password'],
+  }
+  Chef::Log.debug("Post-validate options #{@new_resource.options.inspect}")
 end
 
 action :create do
-  unless exists?
+  unless exists? && node['roles'].include?(new_resource.database_cluster.master_role)
     begin
       db.query("CREATE USER '#{@new_resource.username}'@'#{@new_resource.host}' IDENTIFIED BY '#{@new_resource.password}'")
       @new_resource.updated_by_last_action(true)
@@ -37,7 +43,7 @@ action :create do
 end
 
 action :drop do
-  if exists?
+  if exists? && node['roles'].include?(new_resource.database_cluster.master_role)
     begin
       db.query("DROP USER '#{@new_resource.username}'@'#{@new_resource.host}'")
       @new_resource.updated_by_last_action(true)
@@ -48,15 +54,17 @@ action :drop do
 end
 
 action :grant do
-  begin
-    @new_resource.grant.each do |priv, target|
-      grant_statement = "GRANT #{priv} ON #{@new_resource.target || "*"} TO '#{@new_resource.username}'@'#{@new_resource.host}' IDENTIFIED BY '#{@new_resource.password}'"
-      Chef::Log.info("#{@new_resource}: granting access with statement [#{grant_statement}]")
-      db.query(grant_statement)
-      @new_resource.updated_by_last_action(true)
+  unless node['roles'].include?(new_resource.database_cluster.master_role)
+    begin
+      @new_resource.grant.each do |priv, target|
+        grant_statement = "GRANT #{priv} ON #{@new_resource.target || "*"} TO '#{@new_resource.username}'@'#{@new_resource.host}' IDENTIFIED BY '#{@new_resource.password}'"
+        Chef::Log.info("#{@new_resource}: granting access with statement [#{grant_statement}]")
+        db.query(grant_statement)
+        @new_resource.updated_by_last_action(true)
+      end
+    ensure
+      close
     end
-  ensure
-    close
   end
 end
